# rest-api/public/OpenAPISpec.json
{
	"components": {
...
		"schemas": {
			"IntId": {
				"type": "integer",
				"format": "int32",
				"description": "An integer that serves as a record id. Used for primary keys and foreign keys.",
				"minimum": 1
			},
			"Artist": {
				"description": "Artist",
				"properties": {
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"ArtistId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MediaType": {
				"description": "Media Type",
				"properties": {
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"MediaTypeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Genre": {
				"description": "Genre",
				"properties": {
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"GenreId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Price": {
				"type": "number",
				"format": "float",
				"description": "A price value. Has a precision of 2 significant digits after the decimal point."
			},
...
			"ValidateErrorJSON": {
				"description": "A validation error",
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PlaylistResourceAlbum": {
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Title": {
						"type": "string"
					},
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Artist": {
						"$ref": "#/components/schemas/Artist"
					}
				},
				"required": [
					"AlbumId",
					"Title",
					"ArtistId",
					"Artist"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlaylistResourceTrack": {
				"properties": {
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					},
					"Album": {
						"$ref": "#/components/schemas/PlaylistResourceAlbum"
					},
					"MediaType": {
						"$ref": "#/components/schemas/MediaType"
					},
					"Genre": {
						"$ref": "#/components/schemas/Genre"
					}
				},
				"required": [
					"TrackId",
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice",
					"MediaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlaylistResourcePlaylistTrack": {
				"properties": {
					"PlaylistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Track": {
						"$ref": "#/components/schemas/PlaylistResourceTrack"
					}
				},
				"required": [
					"PlaylistId",
					"TrackId",
					"Track"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlaylistResource": {
				"properties": {
					"PlaylistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"PlaylistTracks": {
						"items": {
							"$ref": "#/components/schemas/PlaylistResourcePlaylistTrack"
						},
						"type": "array"
					}
				},
				"required": [
					"PlaylistId",
					"PlaylistTracks"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TrackResourceAlbum": {
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Title": {
						"type": "string"
					},
					"ArtistId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Artist": {
						"$ref": "#/components/schemas/Artist"
					}
				},
				"required": [
					"AlbumId",
					"Title",
					"ArtistId",
					"Artist"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrackResource": {
				"properties": {
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					},
					"Album": {
						"$ref": "#/components/schemas/TrackResourceAlbum"
					},
					"MediaType": {
						"$ref": "#/components/schemas/MediaType"
					},
					"Genre": {
						"$ref": "#/components/schemas/Genre"
					}
				},
				"required": [
					"TrackId",
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice",
					"MediaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Track": {
				"description": "Track",
				"properties": {
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					}
				},
				"required": [
					"TrackId",
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrackInsertInput": {
				"description": "Track Insert Input",
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					}
				},
				"required": [
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrackUpdateInput": {
				"description": "Track Update Input",
				"properties": {
					"AlbumId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					},
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Composer": {
						"type": "string"
					},
					"Milliseconds": {
						"type": "number",
						"format": "double"
					},
					"Bytes": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					}
				},
				"required": [
					"Name",
					"MediaTypeId",
					"Milliseconds",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "Sample REST API built on the Chinook database, for Hasura demo",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "benoit ranque"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/playlists/{PlaylistId}/Track/{TrackId}": {
			"delete": {
				"operationId": "DeletePlaylistTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlaylistResource"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Playlist"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "PlaylistId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					},
					{
						"in": "path",
						"name": "TrackId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
		"/tracks": {
			"get": {
				"operationId": "GetTracks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrackResource"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Track"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrackInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/tracks/{TrackId}": {
			"get": {
				"operationId": "GetTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrackResource"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "TrackId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Track"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "TrackId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrackUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Track"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "TrackId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}